Experimental hacks to use JDK8. Mac only, no backward compatibity for JDK6.
These changes seem to be sufficient to make JDK8 work.
This module's unit tests pass, and I have evidence that gremlin-node
now accepts references to classes that were compiled for Java 8.

We now have to figure out how to maintain backward compatiblity for JDK6.
Perhaps we can simply detect which JDK is installed in JAVA_HOME and
compile for it?