Somewhat better way to support jdk1.8.
There are still problems we have to solve for a general release,
but this version should do the right thing for jdk1.6 and jdk1.8.

jdk1.7 will work, but it will no longer target 1.6.

In addition to fixing the problem with jdk1.7, we also need to
change the build system so that .java files are compiled to .class
files as part of npm install, instead of checking them in.