Implement testRunner.js.
We need a convenient way to invoke nodeunit multiple
times in independent processes. This testRunner runs
the standard tests in the `test` and `test8` directory
in on process, and then runs each of the files in
testAsyncOptions, each in their own process.

Perhaps I could have done this with a script, but
it's unclear to me whether a shell style script
would work on Windows. I expect this should be portable
across all platforms.